music: address feedback
